training/requirements.txt
snowflake-connector-python
python-dotenv
opencv-python-headless
ultralytics
torch
numpy
pandas
